monster-mash laplace-inflation

+++ src/unique_rows.h
@@ -2,12 +2,12 @@
 #include <eigen3/Eigen/Eigen>
 #include <igl/unique_rows.h>
 
-inline void float_to_int(const Eigen::MatrixXd &data, Eigen::MatrixXi &int_data, int digits = 1)
+inline void float_to_int(const Eigen::MatrixXf &data, Eigen::MatrixXi &int_data, int digits = 1)
 {
     int_data = (data * std::pow(10, digits)).array().round().cast<int>();
 }
 
-inline void hashable_rows(const Eigen::MatrixXd &data, Eigen::MatrixXi &row_index, int digits = 1)
+inline void hashable_rows(const Eigen::MatrixXf &data, Eigen::MatrixXi &row_index, int digits = 1)
 {
     Eigen::MatrixXi int_data;
     float_to_int(data, int_data, digits);
@@ -16,7 +16,7 @@ inline void hashable_rows(const Eigen::MatrixXd &data, Eigen::MatrixXi &row_inde
     igl::unique_rows(int_data, unique_int_data, row_index, IC);
 }
 
-inline void unique_rows(const Eigen::MatrixXd &data, Eigen::MatrixXi &row_index, int digits = 1)
+inline void unique_rows(const Eigen::MatrixXf &data, Eigen::MatrixXi &row_index, int digits = 1)
 {
     // Returns indices of unique rows. It will return the
     // first occurrence of a row that is duplicated:
